♻️ modern Lolliplot
