+++ package.json
@@ -59,7 +59,7 @@
     "react-modal": "1.7.7",
     "react-progress": "0.0.12",
     "react-redux": "5.0.4",
-    "react-relay": "1.0.0",
+    "react-relay": "1.1.0",
     "react-relay-network-layer": "2.0.1",
     "react-router": "4.0.0-beta.4",
     "react-router-dom": "4.0.0-beta.4",
